+++ Float64Int/Float64Int.go
@@ -31,10 +31,10 @@ func Fill(ar []int, newar []KeyVal) []KeyVal {
 	return newar
 }
 
-func Keys(ar []int, newar []KeyVal) []int {
+func Keys(ar []float64, newar []KeyVal) []float64 {
 	l := len(newar)
 	if len(ar) < l {
-		ar = make([]int, l)
+		ar = make([]float64, l)
 	}
 	for i, v := range newar {
 		ar[i] = v.K
