+++ Float64Int32/Float64Int32.go
@@ -31,10 +31,10 @@ func Fill(ar []int32, newar []KeyVal) []KeyVal {
 	return newar
 }
 
-func Keys(ar []int32, newar []KeyVal) []int32 {
+func Keys(ar []float64, newar []KeyVal) []float64 {
 	l := len(newar)
 	if len(ar) < l {
-		ar = make([]int32, l)
+		ar = make([]float64, l)
 	}
 	for i, v := range newar {
 		ar[i] = v.K
