+++ Float64Uint/Float64Uint.go
@@ -31,10 +31,10 @@ func Fill(ar []uint, newar []KeyVal) []KeyVal {
 	return newar
 }
 
-func Keys(ar []uint, newar []KeyVal) []uint {
+func Keys(ar []float64, newar []KeyVal) []float64 {
 	l := len(newar)
 	if len(ar) < l {
-		ar = make([]uint, l)
+		ar = make([]float64, l)
 	}
 	for i, v := range newar {
 		ar[i] = v.K
