+++ Int16Float32/Int16Float32.go
@@ -31,10 +31,10 @@ func Fill(ar []float32, newar []KeyVal) []KeyVal {
 	return newar
 }
 
-func Keys(ar []float32, newar []KeyVal) []float32 {
+func Keys(ar []int16, newar []KeyVal) []int16 {
 	l := len(newar)
 	if len(ar) < l {
-		ar = make([]float32, l)
+		ar = make([]int16, l)
 	}
 	for i, v := range newar {
 		ar[i] = v.K
