+++ Int16Int32/Int16Int32.go
@@ -31,10 +31,10 @@ func Fill(ar []int32, newar []KeyVal) []KeyVal {
 	return newar
 }
 
-func Keys(ar []int32, newar []KeyVal) []int32 {
+func Keys(ar []int16, newar []KeyVal) []int16 {
 	l := len(newar)
 	if len(ar) < l {
-		ar = make([]int32, l)
+		ar = make([]int16, l)
 	}
 	for i, v := range newar {
 		ar[i] = v.K
