+++ Int8Float32/Int8Float32.go
@@ -31,10 +31,10 @@ func Fill(ar []float32, newar []KeyVal) []KeyVal {
 	return newar
 }
 
-func Keys(ar []float32, newar []KeyVal) []float32 {
+func Keys(ar []int8, newar []KeyVal) []int8 {
 	l := len(newar)
 	if len(ar) < l {
-		ar = make([]float32, l)
+		ar = make([]int8, l)
 	}
 	for i, v := range newar {
 		ar[i] = v.K
