+++ Uint16Int8/Uint16Int8.go
@@ -31,10 +31,10 @@ func Fill(ar []int8, newar []KeyVal) []KeyVal {
 	return newar
 }
 
-func Keys(ar []int8, newar []KeyVal) []int8 {
+func Keys(ar []uint16, newar []KeyVal) []uint16 {
 	l := len(newar)
 	if len(ar) < l {
-		ar = make([]int8, l)
+		ar = make([]uint16, l)
 	}
 	for i, v := range newar {
 		ar[i] = v.K
