+++ Uint16Uint/Uint16Uint.go
@@ -31,10 +31,10 @@ func Fill(ar []uint, newar []KeyVal) []KeyVal {
 	return newar
 }
 
-func Keys(ar []uint, newar []KeyVal) []uint {
+func Keys(ar []uint16, newar []KeyVal) []uint16 {
 	l := len(newar)
 	if len(ar) < l {
-		ar = make([]uint, l)
+		ar = make([]uint16, l)
 	}
 	for i, v := range newar {
 		ar[i] = v.K
