+++ Uint32Int8/Uint32Int8.go
@@ -31,10 +31,10 @@ func Fill(ar []int8, newar []KeyVal) []KeyVal {
 	return newar
 }
 
-func Keys(ar []int8, newar []KeyVal) []int8 {
+func Keys(ar []uint32, newar []KeyVal) []uint32 {
 	l := len(newar)
 	if len(ar) < l {
-		ar = make([]int8, l)
+		ar = make([]uint32, l)
 	}
 	for i, v := range newar {
 		ar[i] = v.K
