+++ Uint64Uint/Uint64Uint.go
@@ -31,10 +31,10 @@ func Fill(ar []uint, newar []KeyVal) []KeyVal {
 	return newar
 }
 
-func Keys(ar []uint, newar []KeyVal) []uint {
+func Keys(ar []uint64, newar []KeyVal) []uint64 {
 	l := len(newar)
 	if len(ar) < l {
-		ar = make([]uint, l)
+		ar = make([]uint64, l)
 	}
 	for i, v := range newar {
 		ar[i] = v.K
