+++ Uint8Int/Uint8Int.go
@@ -31,10 +31,10 @@ func Fill(ar []int, newar []KeyVal) []KeyVal {
 	return newar
 }
 
-func Keys(ar []int, newar []KeyVal) []int {
+func Keys(ar []uint8, newar []KeyVal) []uint8 {
 	l := len(newar)
 	if len(ar) < l {
-		ar = make([]int, l)
+		ar = make([]uint8, l)
 	}
 	for i, v := range newar {
 		ar[i] = v.K
