+++ Uint8Int16/Uint8Int16.go
@@ -31,10 +31,10 @@ func Fill(ar []int16, newar []KeyVal) []KeyVal {
 	return newar
 }
 
-func Keys(ar []int16, newar []KeyVal) []int16 {
+func Keys(ar []uint8, newar []KeyVal) []uint8 {
 	l := len(newar)
 	if len(ar) < l {
-		ar = make([]int16, l)
+		ar = make([]uint8, l)
 	}
 	for i, v := range newar {
 		ar[i] = v.K
