+++ UintFloat32/UintFloat32.go
@@ -31,10 +31,10 @@ func Fill(ar []float32, newar []KeyVal) []KeyVal {
 	return newar
 }
 
-func Keys(ar []float32, newar []KeyVal) []float32 {
+func Keys(ar []uint, newar []KeyVal) []uint {
 	l := len(newar)
 	if len(ar) < l {
-		ar = make([]float32, l)
+		ar = make([]uint, l)
 	}
 	for i, v := range newar {
 		ar[i] = v.K
